支持提取协议信息
